+++ services/explorer-api/src/database/controllers/index.ts
@@ -1,6 +1,7 @@
 export * as l2Block from "./l2block/index.js";
 export * as l2Contract from "./l2contract/index.js";
 export * as l2TxEffect from "./l2TxEffect/index.js";
+export * as l2Tx from "./l2Tx/index.js";
 export * as l2 from "./l2/index.js";
 export * as signOfLife from "./sign-of-life.js";
 export * as aztecChainConnection from "./aztec-chain-connection.js";
