+++ services/explorer-api/src/database/controllers/l2Tx/index.ts
@@ -1,1 +1,3 @@
+export * from "./get-tx.js";
+export * from "./replace-txs-with-tx-effects.js";
 export * from "./store.js";
