+++ services/explorer-api/src/database/schema/index.ts
@@ -1,2 +1,3 @@
 export * from "./l2block/index.js";
+export * from "./l2tx/index.js";
 export * from "./l2contract/index.js";
