+++ services/explorer-api/src/database/schema/l2block/body.ts
@@ -27,7 +27,7 @@ export const bodyToTxEffects = pgTable("body_to_tx_effects", {
 export const txEffect = pgTable("tx_effect", {
   hash: varchar("hash").notNull().$type<HexString>().primaryKey(),
   txHash: varchar("hash").notNull().$type<HexString>(),
-  txBirthTime: date("tx_time_of_birth").notNull(),
+  txBirthTime: date("tx_time_of_birth").notNull().defaultNow(),
   addedToChainTime: date("added_to_chain_time").notNull().defaultNow(),
   // TODO: move index to junction table
   index: integer("index").notNull(),
