+++ services/explorer-api/src/event-handler/on-pending-txs.ts
@@ -1,7 +1,7 @@
 import { PendingTxsEvent } from "@chicmoz-pkg/message-registry";
 import { chicmozL2PendingTxSchema } from "@chicmoz-pkg/types";
 import { logger } from "../logger.js";
-import { storeL2Tx } from "../database/controllers/l2contract/store.js";
+import { storeL2Tx } from "../database/controllers/l2Tx/store.js";
 
 export const onPendingTxs = async ({ txs }: PendingTxsEvent) => {
   for (const tx of txs) {
